Rename subpackage zesty registry.
