+++ zesty/utils/rekordo/database.go
@@ -1,4 +1,4 @@
-package utils
+package rekordo
 
 import (
 	"database/sql"
