+++ zesty/utils/rekordo/register.go
@@ -1,4 +1,4 @@
-package utils
+package rekordo
 
 import "sync"
